+++ src/lexer/Lexer.php
@@ -13,40 +13,48 @@ use VovanVE\parser\common\Token;
  *
  * ```php
  * $lexer = (new Lexer)
- *     // all terminals to parse
- *     ->terminals([
- *         // inline tokens literally, order does not matter
+ *     // inline tokens literally
+ *     ->inline([
+ *         // order does not matter
  *         // inline tokens are always hidden
  *         '++',
  *         '+',
- *         '--',
  *         '-',
+ *         '--',
  *         '*',
  *         '/',
- *
- *         // named tokens are RegExp parts
+ *     ])
+ *     // fixed tokens literally
+ *     ->fixed([
+ *         // order does not matter
+ *         'semicolon'    => ';',
+ *         '.colon'       => ':', // hidden named token
+ *         'double_colon' => '::',
+ *         // order between inline and fixed does not matter too
+ *     ])
+ *     // terminals are RegExp parts
+ *     ->terminals([
  *         'int'   => '\\d++',
  *         'const' => '(?&name)',
  *         'var'   => '\\$(?&name)',
- *         '.foo'  => ';',            // hidden named token
+ *         '.foo'  => '\\?++', // hidden named token
  *     ])
  *     // whitespaces and comments to skip completely
  *     ->whitespaces([
  *         '\\s++',           // linear whitespaces
  *         '#\\N*+\\n?+',     // line #comments
  *     ])
- *     // DEFINEs can only be referenced from tokens as named recursion `(?&name)`
+ *     // DEFINEs can only be referenced from tokens and whitespaces as named recursion `(?&name)`
  *     ->defines([
  *         'name' => '[a-z_][a-z_0-9]*+',
  *     ])
  *     ->modifiers('i');
  *
  * Mostly you can define only named tokens. Inline tokens will be added later
- * from grammar inline tokens.
+ * from grammar.
  * ```
  *
  * @package VovanVE\parser
- * @see \VovanVE\parser\LexerBuilder
  */
 class Lexer extends BaseObject
 {
@@ -62,6 +70,16 @@ class Lexer extends BaseObject
      * Key is name and value is a part of RegExp
      */
     private $defines;
+    /**
+     * @var array Fixed terminal definition. Values are plain strings to parse as is.
+     * Keys are names.
+     */
+    private $fixed;
+    /**
+     * @var string[] Inline terminals definition. Values are plain strings to parse as is.
+     * Keys does not matter.
+     */
+    private $inline;
     /**
      * @var array Terminals definition. Items of the array can be
      * either key=>value for named tokens (here Key is name and value is a part of RegExp;
@@ -124,6 +142,8 @@ class Lexer extends BaseObject
         $modifiers = 'u'
     ) {
         $this->defines = $defines;
+        $this->fixed = [];
+        $this->inline = [];
         $this->terminals = $terminals;
         $this->whitespaces = $whitespaces;
         $this->modifiers = $modifiers;
@@ -140,6 +160,8 @@ class Lexer extends BaseObject
 
     /**
      * Create new Lexer extending this one
+     *
+     * > Note: Returned object can be the same one in case of empty additions.
      * @param array $terminals Additional terminals. Both inline and named are acceptable.
      * Duplicating inline tokens is not permitted, but redefinition of named tokens
      * is restricted.
@@ -166,6 +188,12 @@ class Lexer extends BaseObject
 
     /**
      * Create new Lexer extending this one with DEFINEs
+     *
+     * DEFINEs are named regexps to be used as from terminals and whitespaces
+     * with named recursion `(?&name)` to simplify regexp code duplication.
+     * DEFINEs can refer to each other. Order should not matter.
+     *
+     * > Note: Returned object can be the same one in case of empty additions.
      * @param array $defines Additional DEFINEs regexps. Duplicating names is restricted.
      * @return static
      * @since 1.4.0
@@ -190,6 +218,10 @@ class Lexer extends BaseObject
 
     /**
      * Create new Lexer extending this one with whitespaces
+     *
+     * Whitespaces are searched between all actual tokens and completely ignored.
+     *
+     * > Note: Returned object can be the same one in case of empty additions.
      * @param array $whitespaces Additional whitespaces regexps. Duplicating currently
      * is not checked, so it on your own.
      * @return static
@@ -207,7 +239,62 @@ class Lexer extends BaseObject
     }
 
     /**
+     * Create new Lexer extending this one with fixed tokens
+     *
+     * Named tokens defined with fixed strings to parse as is.
+     *
+     * > Note: Returned object can be the same one in case of empty additions.
+     * @param array $fixed Additional fixed tokens. Duplicating names are restricted.
+     * @return static
+     * @throws \InvalidArgumentException In case of name duplication.
+     * @since 1.5.0
+     */
+    public function fixed($fixed)
+    {
+        if (!$fixed) {
+            return $this;
+        }
+
+        $new_fixed = $this->addNamedTokens($this->fixed, $fixed, 'fixed');
+
+        $copy = clone $this;
+        $copy->fixed = $new_fixed;
+        return $copy;
+    }
+
+    /**
+     * Create new Lexer extending this one with inline tokens
+     *
+     * Inline tokens defined only with fixed strings without names. Inline tokens
+     * are always hidden.
+     *
+     * > Note: Returned object can be the same one in case of empty additions.
+     * @param string[] $inline Additional inline tokens. Duplication is not permitted.
+     * @return static
+     * @since 1.5.0
+     */
+    public function inline($inline)
+    {
+        if (!$inline) {
+            return $this;
+        }
+
+        $copy = clone $this;
+        $copy->inline = array_merge($this->inline, $inline);
+        return $copy;
+    }
+
+    /**
      * Create new Lexer extending this one with terminals
+     *
+     * Named tokens defined with regexps. Named regexps from DEFINEs can be
+     * referenced here with named recursion `(?&name)`.
+     *
+     * > Deprecation notice: Inline tokens without names introduced in 1.4.0 should
+     * be defined with `inline()` method since 1.5.0. So, defining inline tokens within
+     * terminals is deprecated.
+     *
+     * > Note: Returned object can be the same one in case of empty additions.
      * @param array $terminals Additional terminals. Both inline and named are acceptable.
      * Duplicating inline tokens is not permitted, but redefinition of named tokens
      * is restricted.
@@ -229,6 +316,14 @@ class Lexer extends BaseObject
 
     /**
      * Create new Lexer extending this one with RegExp modifiers
+     *
+     * Same modifiers will be applied both to tokens and whitespaces regexps.
+     *
+     * Here only "global" modifiers like `u`, `x`, `D`, etc. should be used.
+     * Other modifiers like `i` should (but not required) be used locally
+     * in specific parts like `(?i)[a-z]` or `(?i:[a-z])`.
+     *
+     * > Note: Returned object can be the same one in case of empty additions.
      * @param string $modifiers Additional modifiers to whole regexps.
      * @return static
      * @since 1.4.0
@@ -267,24 +362,50 @@ class Lexer extends BaseObject
 
         $this->checkMapNames($this->defines, self::RE_DEFINE_NAME);
 
+        if (count($this->fixed) !== count(array_unique($this->fixed, SORT_STRING))) {
+            throw new \InvalidArgumentException('Duplicating fixed strings found');
+        }
+
         $inline = [];
+
+        $fixed_hidden = [];
+        $fixed_normal = [];
+        $fixed_map = [];
+        $this->splitTerminals($this->fixed, $inline, $fixed_hidden, $fixed_normal, $fixed_map);
+        if ($inline) {
+            throw new \InvalidArgumentException('Mixing inline tokens in fixed tokens map');
+        }
+
+        $inline = array_combine($this->inline, $this->inline);
+        $more_inline = [];
         $hidden = [];
         $normal = [];
         $map = [];
-        $this->splitTerminals($this->terminals, $inline, $hidden, $normal, $map);
+        $this->splitTerminals($this->terminals, $more_inline, $hidden, $normal, $map);
 
-        $this->checkOverlappedNames($inline, $hidden, $normal);
+        if ($more_inline) {
+            // DEPRECATED: inline tokens in terminals map
+            $inline += $more_inline;
+        }
 
-        $this->hiddens = $hidden;
+        $this->checkOverlappingNames([
+            'fixed hidden' => $fixed_hidden,
+            'fixed normal' => $fixed_normal,
+            'inline' => $inline,
+            'hidden' => $hidden,
+            'normal' => $normal,
+        ]);
+
+        $this->hiddens = $fixed_hidden + $hidden;
         $this->aliased = [];
 
-        $inline_re_map = $this->buildInlines($inline);
-        $same = array_intersect_key($map, $inline_re_map);
+        $fixed_and_inline_re_map = $this->buildFixedAndInlines($fixed_map, $inline);
+        $same = array_intersect_key($map, $fixed_and_inline_re_map);
         if ($same) {
             throw new \LogicException("Duplicating inline and named tokens: " . join(', ', $same));
         }
 
-        $terminals_map = $inline_re_map + $map;
+        $terminals_map = $fixed_and_inline_re_map + $map;
         if (!$terminals_map) {
             throw new \InvalidArgumentException('No terminals defined');
         }
@@ -385,6 +506,27 @@ class Lexer extends BaseObject
     }
 
     /**
+     * Extends array of named tokens
+     * @param array $oldTokens Existing tokens
+     * @param array $addTokens New tokens to add
+     * @param string $errorType Tokens type to insert in error message
+     * @return array New merged array of tokens
+     * @throws \InvalidArgumentException In case of name duplication.
+     * @since 1.5.0
+     */
+    private function addNamedTokens($oldTokens, $addTokens, $errorType)
+    {
+        $dup_keys = array_intersect_key($oldTokens, $addTokens);
+        if ($dup_keys) {
+            throw new \InvalidArgumentException(
+                "Cannot redefine $errorType: " . var_export($dup_keys, true)
+            );
+        }
+
+        return $oldTokens + $addTokens;
+    }
+
+    /**
      * Extends array of mixed tokens definition with another array of tokens
      *
      * Array of mixed tokens has both named (name=>definition) and anonymous
@@ -446,31 +588,28 @@ class Lexer extends BaseObject
     /**
      * Check defined tokens for duplicating names
      *
-     * Different tokens must to produce different output Tokens.
-     * So names and inline tokens must not overlap.
-     * @param array $inline Extracted inline tokens `["plain" => mixed, ...]`
-     * @param array $hidden Extracted hidden tokens `["name" => mixed, ...]`
-     * @param array $normal Extracted normal tokens `["name" => mixed, ...]`
+     * Each token's name must be used only once. Inline tokens must not overlap
+     * with names of named tokens.
+     * @param array $maps Map on different tokens' maps. Key is token type for error message.
+     * Value is tokens map of the type in form `[name => mixed, ...]`. Inline tokens use
+     * its strings as name.
      * @throws \InvalidArgumentException Some overlapping names
-     * @since 1.4.0
+     * @since 1.5.0
      */
-    private function checkOverlappedNames(array $inline, array $hidden, array $normal)
+    private function checkOverlappingNames(array $maps)
     {
-        foreach (
-            [
-                'named normal and hidden' => [$normal, $hidden],
-                'named tokens and inline quoted' => [$normal, $inline],
-                'named hidden tokens and inline quoted' => [$hidden, $inline],
-            ]
-            as $message => $pair
-        ) {
-            list ($a, $b) = $pair;
-            $same = array_intersect_key($a, $b);
-            if ($same) {
-                throw new \InvalidArgumentException(
-                    "Duplicating $message tokens: " . join(', ', $same)
-                );
+        $index = 0;
+        foreach ($maps as $type => $map) {
+            $rest_maps = array_slice($maps, $index + 1, null, true);
+            foreach ($rest_maps as $type2 => $map2) {
+                $same = array_intersect_key($map, $map2);
+                if ($same) {
+                    throw new \InvalidArgumentException(
+                        "Duplicating $type tokens and $type2 tokens: " . join(', ', $same)
+                    );
+                }
             }
+            ++$index;
         }
     }
 
@@ -482,27 +621,42 @@ class Lexer extends BaseObject
      * Inline tokens all are aliased with generated names. Returned regexps map has that
      * generated named in keys. Aliases map is stored in `$aliased` property. As of all inline
      * tokens are hidden, `$hiddens` map will be updated too with generated names.
+     * @param array $fixedMap Fixed tokens map
      * @param array $inlines List of inline token texts.
      * @return array Returns regexps map with generated names in keys. Also properties `$aliased`
      * and `$hiddens` will be updated.
-     * @since 1.4.0
+     * @since 1.5.0
      */
-    private function buildInlines(array $inlines)
+    private function buildFixedAndInlines(array $fixedMap, array $inlines)
     {
-        // sort in reverse order to let more long items with match first
+        $overlapped = array_intersect($fixedMap, $inlines);
+        if ($overlapped) {
+            throw new \InvalidArgumentException(
+                "Duplicating fixed tokens and inline tokens strings: "
+                . join(', ', array_map('json_encode', $overlapped))
+            );
+        }
+
+        // $fixedMap all keys are valid names, so + with integer keys will not loss anything
+        $all = $fixedMap + array_values($inlines);
+
+        // sort in reverse order to let more long items match first
         // so /'$$' | '$'/ will find ['$$', '$'] in '$$$' and not ['$', '$', '$']
-        rsort($inlines, SORT_STRING);
+        arsort($all, SORT_STRING);
 
         $re_map = [];
 
         $alias_name = 'a';
-        foreach ($inlines as $text) {
-            $name = '_' . $alias_name;
-            // string increment
-            $alias_name++;
+        foreach ($all as $name => $text) {
+            // inline?
+            if (is_int($name)) {
+                $name = '_' . $alias_name;
+                // string increment
+                $alias_name++;
 
-            $this->aliased[$name] = $text;
-            $this->hiddens[$text] = true;
+                $this->aliased[$name] = $text;
+                $this->hiddens[$text] = true;
+            }
             $re_map[$name] = preg_quote($text, '/');
         }